requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
Eroseway==0.0.3
idna==3.7
MarkupSafe==2.1.5
mysqlclient==2.2.4
pyactiveresource==2.2.2
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
ShopifyAPI==12.6.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
